wcnss_service: Check for modem support at the run time
Some of the targets do not support modem. While starting wcnss
service it would check if the modem initialization is successful.
Only if it is successful, wcnss service is started. If not, this
would introduce a delay in the of process of driver load. To
address this, check if the target has the modem support.

CRs-Fixed: 715392
Change-Id: Ib8ab8718f11389c6e540606a82de87ce4c849307