Adding support for passing userdata in nan_register_handler
This commit is to enhance nan_register_handler to accept
a userdata pointer which will be returned back as part of the
nan response or nan event. This can be used by the upper layer
to pass a context or object pointer which it can get back when
the callback from nan is called.

Change-Id: I572a3f094226f4f5fd77530745451d65c20a892b
CRs-Fixed: 769281