Wifi-HAL: Call notifyResponse when link stats get success
When link stats get success, notifyResponse supposed to be called
to indicate stats to framework, and as part of it, clearStats will
free the memory of the stats.

However in current code, notifyResponse is not called, which makes
memory of the stats not freed (by clearStats).
Correct it to call notifyResponse in such case to avoid memory leak.

Change-Id: Iea72dfdf9936da8d6bb02c05535957ff590fb86b
CRs-Fixed: 2169626